split metrics from collation generation
